phase-2/requirements.txt
# Core
streamlit==1.35.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
slowapi==0.1.9

# ML
scikit-learn==1.4.2
lightgbm==4.3.0
catboost==1.2.5
shap==0.45.0
joblib==1.4.2
imbalanced-learn==0.12.3

# Data
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# Agent
langgraph==0.2.6
langchain==0.2.6
langchain-groq==0.1.6
langchain-community==0.2.6

# RAG
faiss-cpu==1.8.0
sentence-transformers==3.0.1


# HTTP
httpx==0.27.0
requests==2.32.3
grpcio==1.62.2

# Viz
plotly==5.22.0
matplotlib==3.9.0
